+++ DevFreela.API/Models/Input/CreateProjectCommentInputModel.cs
@@ -5,10 +5,10 @@ namespace DevFreela.API.Models.Input
     public class CreateProjectCommentInputModel
     {
         public string Content { get; set; }
-        public int ProjectId { get; private set; }
-        public int UserId { get; set; }
+        public long ProjectId { get; private set; }
+        public long UserId { get; set; }
 
-        public void SetProjectId(int id) => ProjectId = id;
+        public void SetProjectId(long id) => ProjectId = id;
 
         public ProjectComment ToEntity()
             => new(Content, ProjectId, UserId);
